rename lint check script to style check to keep it unified across skeletons, add shellcheck to style-check script, fix shellcheck warnings

+++ clear-cache.sh
@@ -1,14 +1,13 @@
 #!/bin/sh -e
 
-SCRIPT_DIR=$(cd $(dirname ${0}); pwd)
-
+DIR=$(dirname "${0}")
+SCRIPT_DIR=$(cd "${DIR}"; pwd)
 echo "Deleting cache files."
-
 FILES="build  .coverage .sonar"
 
 for FILE in ${FILES}; do
     if [ -e "${FILE}" ]; then
         echo "rm -rf ${FILE}"
-        rm -rf "${SCRIPT_DIR}/${FILE}"
+        rm -rf "${SCRIPT_DIR:?}/${FILE}"
     fi
 done
